Added support for decimal numbers

+++ fs-extras/Extra008/Program.cs
@@ -1,14 +1,14 @@
 ﻿using System;
 
 Console.Write("How much money do you have?: ");
-int moneys = int.Parse(Console.ReadLine());
+double moneys = double.Parse(Console.ReadLine());
 Console.Write("Price of food: ");
-int price = int.Parse(Console.ReadLine());
+double price = double.Parse(Console.ReadLine());
 string response = (enoughMoney(moneys, price)) ? "Money left: " + Convert.ToString(moneys - price) : "Paastoamaan!";
 
 Console.WriteLine($"{response}");
 
-bool enoughMoney(int moneys, int price)
+bool enoughMoney(double moneys, double price)
 {
     return moneys >= price;
 }
